DEV V20151226.3 Programming on FGTracker service

+++ service/update.php
@@ -4,6 +4,30 @@ class UpdateMgr
 	
 	public function __construct () 
 	{}
+	
+	function fgt_pg_query_params($sql,$sql_parm)
+	{	/*if $sql_parm = NULL, pg_query is used
+		  if $sql_parm is an array, pg_query_params is used
+		*/
+		global $fgt_error_report,$var,$fgt_sql;
+		if($sql_parm==NULL)
+			$res=pg_query($fgt_sql->conn,$sql);
+		else
+			$res=pg_query_params($fgt_sql->conn,$sql,$sql_parm);
+		if ($res===false or $res==NULL)
+		{
+			$phpErr=error_get_last();
+			$message="Internal DB Error - ".pg_last_error ($fgt_sql->conn);
+			$fgt_error_report->fgt_set_error_report("U_EFT",$message,E_ERROR);
+			$message="SQL command of last error: ".$sql;
+			$fgt_error_report->fgt_set_error_report("U_EFT",$message,E_ERROR);
+			$message="PHP feedback of last error: ".$phpErr['message'];
+			$fgt_error_report->fgt_set_error_report("U_EFT",$message,E_ERROR);
+			$var['exitflag']=true;
+			return false;
+		}return $res;
+	}
+	
 	public function fix_erric_data()
 	{
 		global $fgt_sql,$fgt_error_report;
@@ -13,9 +37,9 @@ class UpdateMgr
 		
 		/*Waypoints with altitude < -9000*/
 		$sql="delete from waypoints where altitude<-9000";
-		$res=pg_query($fgt_sql->conn,$sql);
-		if ($res===false)
-			return 1;
+		$res=$this->fgt_pg_query_params($sql,Array());
+		if($res===false)
+			return;
 		$nr=pg_affected_rows($res);
 		pg_free_result($res);
 		$message="$nr waypoints with altitude < -9000 deleted.";
@@ -23,7 +47,9 @@ class UpdateMgr
 		
 		/*Flights with negative flight duration*/
 		$sql="delete from flights where start_time > end_time";
-		$res=pg_query($fgt_sql->conn,$sql);
+		$res=$this->fgt_pg_query_params($sql,Array());
+		if($res===false)
+			return;
 		$nr=0;
 		$nr=pg_affected_rows($res);
 		pg_free_result($res);
@@ -42,7 +68,9 @@ class UpdateMgr
 		
 		$flight_array=Array();
 		$sql="SELECT id,callsign FROM flights_all WHERE status='CLOSED' AND (effective_flight_time IS NULL)";
-		$res=pg_query($fgt_sql->conn,$sql);
+		$res=$this->fgt_pg_query_params($sql,Array());
+		if($res===false)
+			return;
 		$nr=pg_num_rows($res);
 		$message="$nr flights need to be updated";
 		$fgt_error_report->fgt_set_error_report("U_EFT",$message,E_NOTICE);
@@ -58,7 +86,9 @@ class UpdateMgr
 			if($var['exitflag']===true)
 				return;
 			$sql="SELECT EXTRACT(EPOCH FROM time) AS time,longitude,latitude,altitude FROM waypoints_all WHERE flight_id=$flight_id AND (longitude!=0 OR latitude!=0 OR altitude!=0) AND altitude>=-9000 ORDER BY time";
-			$res=pg_query($fgt_sql->conn,$sql);
+			$res=$this->fgt_pg_query_params($sql,Array());
+			if($res===false)
+				return;
 			$nr=pg_num_rows($res);
 
 			$array=Array();
@@ -106,30 +136,27 @@ class UpdateMgr
 				
 			if ($nr<1)
 			{
-				//$query.="UPDATE flights set effective_flight_time=0, start_icao=NULL,end_icao=NULL where id=$flight_id;";
-				//$query.="UPDATE flights_archive set effective_flight_time=0, start_icao=NULL,end_icao=NULL where id=$flight_id;";
+				$query.="UPDATE flights set effective_flight_time=0, start_icao=NULL,end_icao=NULL where id=$flight_id;";
+				$query.="UPDATE flights_archive set effective_flight_time=0, start_icao=NULL,end_icao=NULL where id=$flight_id;";
 
 			}else
 			{
-				//$query.="UPDATE flights set effective_flight_time=$effectiveFlightTime, start_icao='$dep_airport[0]',end_icao='$arr_airport[0]' where id=$flight_id;";
-				//$query.="UPDATE flights_archive set effective_flight_time=$effectiveFlightTime, start_icao='$dep_airport[0]',end_icao='$arr_airport[0]' where id=$flight_id;";
+				$query.="UPDATE flights set effective_flight_time=$effectiveFlightTime, start_icao='$dep_airport[0]',end_icao='$arr_airport[0]' where id=$flight_id;";
+				$query.="UPDATE flights_archive set effective_flight_time=$effectiveFlightTime, start_icao='$dep_airport[0]',end_icao='$arr_airport[0]' where id=$flight_id;";
 			}
 			
 			if ($j%100==0)
 			{
 				$message="COMMIT $j done";
 				$fgt_error_report->fgt_set_error_report("U_EFT",$message,E_NOTICE);
-				$res=pg_query($fgt_sql->conn,$query);
-				pg_free_result($res);
+				if($this->fgt_pg_query_params($query,NULL)===false)
+					break;
 				$query="";
 			}
 			$j++;
 		}
 		if ($query!="")
-		{
-				$res=pg_query($fgt_sql->conn,$query);
-				pg_free_result($res);
-		}
+			$this->fgt_pg_query_params($query,NULL);
 		$message="$j flights updated";
 		$fgt_error_report->fgt_set_error_report("U_EFT",$message,E_WARNING);
 	}
@@ -146,7 +173,9 @@ class UpdateMgr
 		$sql.="select setval('temp_cache_top100_alltime_rank_seq',1);";
 		$sql.="INSERT INTO $temp SELECT f.callsign AS callsign,justify_hours(sum(f.end_time-f.start_time)) AS flighttime,justify_hours(sum(effective_flight_time)* '1 second'::interval) AS effective_flight_time FROM flights_all as f GROUP BY f.callsign HAVING sum(effective_flight_time) is not null ORDER BY sum(effective_flight_time) DESC;";
 		$sql.="update $temp set rank=rank-1;";
-		$res=pg_query($fgt_sql->conn,$sql);
+		$res=$this->fgt_pg_query_params($sql,Array());
+		if($res===false)
+			return;
 		
 		$sql="Truncate table $perm;";
 		$sql.="INSERT INTO $perm select callsign,flighttime,rank,null,null,effective_flight_time from $temp;";
@@ -154,7 +183,9 @@ class UpdateMgr
 		$sql.="update $perm AS P set effective_lastweek=(select justify_hours(sum(effective_flight_time)* '1 second'::interval) AS flighttime FROM flights as f WHERE (age(now(),f.end_time)<='7 days'::interval) and f.callsign=P.callsign HAVING sum(f.end_time-f.start_time)>'00:00:05'::interval);";
 		$sql.="update $perm AS P set last30days=(select justify_hours(sum(f.end_time-f.start_time)) AS flighttime FROM flights as f WHERE (age(now(),f.end_time)<='30 days'::interval) and f.callsign=P.callsign HAVING sum(f.end_time-f.start_time)>'00:00:05'::interval);";
 		$sql.="update $perm AS P set effective_last30days=(select justify_hours(sum(effective_flight_time)* '1 second'::interval) AS flighttime FROM flights as f WHERE (age(now(),f.end_time)<='30 days'::interval) and f.callsign=P.callsign HAVING sum(f.end_time-f.start_time)>'00:00:05'::interval);";
-		$res=pg_query($fgt_sql->conn,$sql);
+		$res=$this->fgt_pg_query_params($sql,Array());
+		if($res===false)
+			return;
 	}
 }
 ?>
